Improve initial window size
It appears we must first set the font and then "show" the widgets. This
results in the correct window size -- unless you're using an internal
border size other than "1px". If you do, the initial window size might
be off by some characters.

1px is VTE's default value. config.def.h has been adjusted accordingly.

This is a follow-up to a0e90d9.

Yes, this is still not a 100% satisfying solution. However, xiate is
meant to be used with tiling window managers and this issue is almost
completely irrelevant when using a tiler. Hence, it's a "90% solution".