Fix setting window titles
Whoops. We have connected to the wrong signal: When we're calling
gtk_window_set_title(), then we must listen for "window-title-changed"
signals.

This just happened to work most of the time because there is a common
escape sequence that sets both icon name and window title. Not all
programs use that sequence, though. Some only set the window title and
don't touch the icon name.

Note: As of this commit, we officially only set the window title.
Changing icon names is not supported.

Why is that?

GTK provides no way in GtkWindow to set the icon name. We could set icon
names by using GDK directly. I think, though, that we then end up with
some race conditions (due to windows being realized or not). I'd rather
avoid that.

However, GTK chooses to also set the icon name when changing the window
title. I can only assume they are doing this due to some usability
guidelines. I say this is a GTK issue. xiate never calls anything
related to icon names, so GTK shouldn't set the icon name. If GTK still
insists on doing so, then so be it.