daemon: Keep socket open until window is closed
A client shall be able to tell when the window has been closed. This
brings back the usual semantics of other terminals, so you can do
something like this in a script:

    #!/bin/sh

    xiatec -e vim /tmp/file
    grep /tmp/file something

That "grep" shall only be executed once the terminal window running
"vim" has been closed.

This commit changes the protocol between server and client:

    - The "H" command needs a NUL byte following it.
    - The entire command sequence must be enclosed by "S\000 .. F\000".