client: Reimplement in C, use new protocol
An earlier version of this branch kept the shell script and kept using
socat. This added considerable overhead (startup time and memory
footprint), and it was harder to understand what's going on since it
required some tricks.

Let's end the era of a shell client (as ccrusius on GitHub already
suggested almost a year ago).