xiatec: Exit with 0 or 1
There is no point in sending the original exit code. In the parent
process running xiatec, you can never tell whether it was an internal
xiate error or whether the child program had a problem.

The only thing you can do is signal "success" or "failure" to xiatec's
parent. And this is what we do now.

(Yes, you could think of using one fixed exit code to signal "internal
xiate error", but what if your terminal window runs another command like
ssh that tries to do the same thing? This problem can't be solved in a
clean way.)

Note that 0 or 1 is still good enough for most use cases and it's much
better than some other terminals which completely ignore the child's
exit code.